moving functions to libs
